Title: 토끼와 경주, Time: 23ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 241004/토끼와 경주/rabit-and-race.cpp
@@ -46,7 +46,7 @@ int ans;
 
 priority_queue<Rabbit> pq;
 
-int allScore = 0;
+long long allScore = 0;
 unordered_map<int, int> minusScore;
 unordered_map<int, int> rabbitDistance;
 int minMInusScore = 0;
